+++ CHANGELOG.md
@@ -23,6 +23,15 @@ Classify the change according to the following categories:
     ### Deprecated
     ### Removed
 
+## Develop degradation-cleanup
+### Added
+- Battery residual value if choosing replacement strategy for degradation
+### Changed
+- Revised the battery degradation model, refactoring some methods to increase model-building efficiency and reformulating indicator constraints as big-M constraints with smaller big-M's to reduce solve time.
+- Edited several documentation entries and docstrings for clarity.
+### Removed
+- 80% scaling of battery maintenance costs when using augmentation strategy
+
 ## v0.48.0
 ### Added
 - Added new file `src/core/ASHP.jl` with new technology **ASHP**, which uses electricity as input and provides heating and/or cooling as output; load balancing and technology-specific constraints have been updated and added accordingly